fix level editor
